requirements.txt
Django
django-modeltranslation
django-ckeditor-5
pillow
django-colorfield